requirements.txt
langgraph
google-genai
ollama
python-dotenv
GitPython
supabase
pydriller
radon
requests
typing-extensions
